fix point data type and new point calculate function

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->integer('level')->default(4);
-            $table->unsignedInteger('point')->default('1')->nullable();
+            $table->bigInteger('point')->default('1')->nullable();
             $table->rememberToken();
             $table->timestamps();
         });
